refactor: rename folder for catalog's filters files
