+++ src/app/view/main/catalog/catalog-view.ts
@@ -5,7 +5,7 @@ import { CATALOG_CLASSES, CATALOG_TEXT } from './catalog-view-types';
 import { ProductAPI } from '../../../../api/product-api/product-api';
 import { ProductCard, ProductCards } from '../../../../api/product-api/product-api-types';
 import { clearElement } from '../../../utils/clear-element';
-import { FiltersView } from './refine/filters-view';
+import { FiltersView } from './filters/filters-view';
 
 class CatalogView extends View {
   private content: ElementCreator | null;
