+++ src/app/view/main/catalog/filters/filters-view.ts
